move QgsPolygon tests from QgsGeometry
